refactor: move Firestore query logic into Pinia store for better state management
